refactor: TableGroupService의 검증 로직을 TableGroupExternalValidator로 옮기고 도메인에서 검증하도록 변경
