+++ src/main/java/kitchenpos/tablegroup/application/TableGroupService.java
@@ -1,32 +1,30 @@
 package kitchenpos.tablegroup.application;
 
 import java.util.List;
-import java.util.stream.Collectors;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import kitchenpos.order.application.OrderService;
 import kitchenpos.ordertable.domain.OrderTable;
 import kitchenpos.ordertable.domain.OrderTableRepository;
 import kitchenpos.ordertable.exception.NotFoundOrderTableException;
 import kitchenpos.tablegroup.domain.TableGroup;
+import kitchenpos.tablegroup.domain.TableGroupExternalValidator;
 import kitchenpos.tablegroup.domain.TableGroupRepository;
 import kitchenpos.tablegroup.dto.TableGroupAddRequest;
 import kitchenpos.tablegroup.dto.TableGroupResponse;
-import kitchenpos.tablegroup.exception.CanNotUngroupByOrderStatusException;
 import kitchenpos.tablegroup.exception.NotFoundTableGroupException;
 
 @Service
 public class TableGroupService {
 
-	private final OrderService orderService;
+	private final TableGroupExternalValidator tableGroupExternalValidator;
 	private final OrderTableRepository orderTableRepository;
 	private final TableGroupRepository tableGroupRepository;
 
-	public TableGroupService(final OrderService orderService, final OrderTableRepository orderTableRepository,
-		final TableGroupRepository tableGroupRepository) {
-		this.orderService = orderService;
+	public TableGroupService(TableGroupExternalValidator tableGroupExternalValidator,
+		OrderTableRepository orderTableRepository, TableGroupRepository tableGroupRepository) {
+		this.tableGroupExternalValidator = tableGroupExternalValidator;
 		this.orderTableRepository = orderTableRepository;
 		this.tableGroupRepository = tableGroupRepository;
 	}
@@ -51,22 +49,11 @@ public class TableGroupService {
 	@Transactional
 	public void ungroup(final Long tableGroupId) {
 		final TableGroup tableGroup = findTableGroup(tableGroupId);
-		validateUngroup(tableGroup.getOrderTables());
-		tableGroup.ungroup();
+		tableGroup.ungroup(tableGroupExternalValidator);
 	}
 
 	private TableGroup findTableGroup(Long id) {
 		return tableGroupRepository.findById(id)
 			.orElseThrow(NotFoundTableGroupException::new);
-	}
-
-	private void validateUngroup(final List<OrderTable> orderTables) {
-		final List<Long> orderTableIds = orderTables.stream()
-			.map(OrderTable::getId)
-			.collect(Collectors.toList());
-
-		if (orderService.existsOrderStatusCookingOrMeal(orderTableIds)) {
-			throw new CanNotUngroupByOrderStatusException();
-		}
 	}
 }
