+++ src/test/java/kitchenpos/tablegroup/application/TableGroupServiceTest.java
@@ -16,11 +16,11 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
-import kitchenpos.order.application.OrderService;
 import kitchenpos.ordertable.domain.OrderTable;
 import kitchenpos.ordertable.domain.OrderTableRepository;
 import kitchenpos.ordertable.exception.NotFoundOrderTableException;
 import kitchenpos.tablegroup.domain.TableGroup;
+import kitchenpos.tablegroup.domain.TableGroupExternalValidator;
 import kitchenpos.tablegroup.domain.TableGroupRepository;
 import kitchenpos.tablegroup.dto.TableGroupAddRequest;
 import kitchenpos.tablegroup.dto.TableGroupResponse;
@@ -33,7 +33,7 @@ class TableGroupServiceTest {
 	private TableGroupService tableGroupService;
 
 	@Mock
-	private OrderService orderService;
+	private TableGroupExternalValidator tableGroupExternalValidator;
 	@Mock
 	private OrderTableRepository orderTableRepository;
 	@Mock
@@ -87,7 +87,6 @@ class TableGroupServiceTest {
 		final TableGroup 그룹 = TableGroup.of(1L, Arrays.asList(주문테이블1, 주문테이블2));
 
 		given(tableGroupRepository.findById(any())).willReturn(Optional.of(그룹));
-		given(orderService.existsOrderStatusCookingOrMeal(anyList())).willReturn(false);
 
 		tableGroupService.ungroup(그룹.getId());
 
@@ -95,19 +94,5 @@ class TableGroupServiceTest {
 			() -> assertThat(주문테이블1.getTableGroupId()).isNull(),
 			() -> assertThat(주문테이블2.getTableGroupId()).isNull()
 		);
-	}
-
-	@DisplayName("그룹 해지: 주문 테이블의 주문 상태가 조리 혹은 식사 상태이면 예외발생")
-	@Test
-	void ungroup_order_table_status_cooking_or_meal() {
-		final OrderTable 주문테이블1 = OrderTable.of(1L, null, 4, true);
-		final OrderTable 주문테이블2 = OrderTable.of(2L, null, 3, true);
-		final TableGroup 그룹 = TableGroup.of(1L, Arrays.asList(주문테이블1, 주문테이블2));
-
-		given(tableGroupRepository.findById(any())).willReturn(Optional.of(그룹));
-		given(orderService.existsOrderStatusCookingOrMeal(anyList())).willReturn(true);
-
-		assertThatExceptionOfType(CanNotUngroupByOrderStatusException.class)
-			.isThrownBy(() -> tableGroupService.ungroup(그룹.getId()));
 	}
 }
