+++ src/test/java/kitchenpos/tablegroup/domain/TableGroupTest.java
@@ -2,6 +2,7 @@ package kitchenpos.tablegroup.domain;
 
 import static org.assertj.core.api.Assertions.*;
 import static org.junit.jupiter.api.Assertions.*;
+import static org.mockito.Mockito.*;
 
 import java.util.Arrays;
 import java.util.Collections;
@@ -10,6 +11,7 @@ import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import kitchenpos.order.application.OrderService;
 import kitchenpos.ordertable.domain.OrderTable;
 import kitchenpos.tablegroup.exception.CanNotGroupByEmptyException;
 import kitchenpos.tablegroup.exception.CanNotGroupByGroupingAlreadyException;
@@ -83,11 +85,12 @@ class TableGroupTest {
 	@DisplayName("그룹해지")
 	@Test
 	void ungroup() {
+		final TableGroupExternalValidator externalValidator = new TableGroupExternalValidator(mock(OrderService.class));
 		final OrderTable 그룹_빈테이블1 = OrderTable.of(1L, null, 1, true);
 		final OrderTable 그룹_빈테이블2 = OrderTable.of(2L, null, 2, true);
 		final TableGroup 그룹 = TableGroup.of(Arrays.asList(그룹_빈테이블1, 그룹_빈테이블2));
 
-		그룹.ungroup();
+		그룹.ungroup(externalValidator);
 
 		assertAll(
 			() -> assertThat(그룹_빈테이블1.getTableGroupId()).isNull(),
